refactored + forecast code
finished the forecast contoller and service. creted the forecasat roustes. But most importantly refactored the National weather service api call. set iutto operate as a function along with 3 other helper/callback functions. it currently hit the location id roue andd gets all location id info. the cycles through each on for and plugs it into the api url for the national weather service. it then takes the forcast info it gets from that call and stores it in a variable then loops throughthat info and configures it for our foecast api. and finally it then goes through and stores it via POST req to our api. Wheewwhh lol. all route have been tested and run. I will now add in chron job and the delete call.--'